pyproject.toml
[tool.poetry]
name = "kundli-calculator"
version = "0.1.0"
description = "South Indian style Kundli calculation web service"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
sqlalchemy = "^1.4.23"
pyswisseph = "^2.8.0"
pandas = "^1.3.0"
uvicorn = "^0.15.0"
python-multipart = "^0.0.5"
pydantic = "^1.8.2"
alembic = "^1.7.1"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
